feature:Implement the function of delete users

+++ traffic-system/src/main/java/com/zdf/trafficsystem/user/repo/UserRepository.java
@@ -5,8 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 /**
+ * @return null
  * @author mrzhang
+ * @description TODO
+ * @date 2024/3/16 19:43
  */
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 }
